+++ xenian_channel/bot/models/channelsettings.py
@@ -1,7 +1,8 @@
+from contextlib import contextmanager
 from threading import Lock
 
 from elasticsearch.exceptions import ConnectionError, NotFoundError
-from mongoengine import Document, DynamicField, ListField, ReferenceField, StringField
+from mongoengine import DictField, Document, ListField, ReferenceField, StringField
 from urllib3.exceptions import NewConnectionError
 
 from xenian_channel.bot.models.tg_chat import TgChat
@@ -23,26 +24,35 @@ class ChannelSettings(Document):
     import_messages = ListField(ReferenceField(TgMessage))
 
     # should actually be DictField(ListField(ReferenceField(TgMessage))) but it has errors if used like so
-    queued_messages = DynamicField(default={})
-    import_messages_queue = DynamicField(default={})
+    queued_messages = DictField(default={})
+    import_messages_queue = DictField(default={})
+    scheduled_messages = DictField(default={})
 
     save_lock = Lock()
 
     def __repr__(self):
         return f'{str(self.user)} - {str(self.chat)}'
 
-    def save(self, *args, **kwargs):
+    @contextmanager
+    def save_contextmanager(self, *args, **kwargs):
         try:
             self.save_lock.acquire()
-            try:
-                if hasattr(self, '_changed_fields') and 'sent_messages' in self._changed_fields:
-                    before = self._get_collection().find_one(({'_id': self.pk}))
-                    newly_sent = filter(lambda item: item.message_id not in before['sent_messages'], self.sent_messages)
-                    for message in newly_sent:
-                        message.add_to_image_match(metadata={'chat_id': self.chat.id})
-            except (ConnectionError, NewConnectionError, NotFoundError):
-                pass
-
+            yield
+            self.before_save()
             super().save(*args, **kwargs)
         finally:
             self.save_lock.release()
+
+    def before_save(self):
+        try:
+            if hasattr(self, '_changed_fields') and 'sent_messages' in self._changed_fields:
+                before = self._get_collection().find_one(({'_id': self.pk}))
+                newly_sent = filter(lambda item: item.message_id not in before['sent_messages'], self.sent_messages)
+                for message in newly_sent:
+                    message.add_to_image_match(metadata={'chat_id': self.chat.id})
+        except (ConnectionError, NewConnectionError, NotFoundError):
+            pass
+
+    def save(self, *args, **kwargs):
+        with self.save_contextmanager():
+            pass